Basic Form that generates sorted random numbers.
The script sorts the numbers with two nested For loops, an Array for the random numbers, and a sorted List for the results. The Windows Form has options for the amount of numbers as well as the higher/lower limits for those numbers. Two ways that come to mind for how I could improve this project would be: By moving the main parts out into subroutines and functions, and by allowing for output to a text file. I'd do those things, but this was already a 30-minute project I spent an extra couple hours on.